Refactor student login  for readability and fix Django 1.11 issues
