+++ cms/envs/test.py
@@ -39,16 +39,9 @@ from lms.envs.test import (
     REGISTRATION_EXTRA_FIELDS,
 )
 
-# Add some host names used in assorted tests
+# Allow all hosts during tests, we use a lot of different ones all over the codebase.
 ALLOWED_HOSTS = [
-    'localhost',
-    'logistration.testserver',
-    '.testserver.fake',
-    'test-site.testserver',
-    'testserver.fakeother',
-    'edx.org',
-    'microsite.example.com',
-    'testserver2',
+    '*'
 ]
 
 # mongo connection settings
