+++ common/djangoapps/student/tests/test_recent_enrollments.py
@@ -18,7 +18,8 @@ from openedx.core.djangoapps.site_configuration.tests.test_util import with_site
 from shoppingcart.models import DonationConfiguration
 from student.models import CourseEnrollment, DashboardConfiguration
 from student.tests.factories import UserFactory
-from student.views import _get_recently_enrolled_courses, get_course_enrollments
+from student.views import get_course_enrollments
+from student.views.dashboard import _get_recently_enrolled_courses
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory
