+++ common/djangoapps/student/views/__init__.py
@@ -1,0 +1,8 @@
+"""
+Combines all of the broken out student views
+"""
+
+# pylint: disable=wildcard-import
+from dashboard import *
+from login import *
+from management import *
