+++ openedx/core/djangoapps/external_auth/tests/test_shib.py
@@ -322,7 +322,7 @@ class ShibSPTest(CacheIsolationTestCase):
                     'terms_of_service': u'true',
                     'honor_code': u'true'}
 
-        with patch('student.views.AUDIT_LOG') as mock_audit_log:
+        with patch('student.views.management.AUDIT_LOG') as mock_audit_log:
             self.client.post('/create_account', data=postvars)
 
         mail = identity.get('mail')
