+++ pavelib/quality.py
@@ -604,7 +604,8 @@ def _get_count_from_last_line(filename, file_type):
     It is returning only the value (as a floating number).
     """
     last_line = _get_report_contents(filename, last_line_only=True).strip()
-    if file_type is "python_complexity":
+
+    if file_type == "python_complexity":
         # Example of the last line of a complexity report: "Average complexity: A (1.93953443446)"
         regex = r'\d+.\d+'
     else:
